requirements.txt
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
mss==9.0.1
keyboard==0.13.5
google-generativeai==0.5.2
pyaudiowpatch
SpeechRecognition==3.10.3
Pillow==10.3.0
dotenv
